remove serverOnly
As discussed in discord,  removing this flag.
Changing this flag at runtime does not work at all.
and it is also bugged,  see #389

The use case this addresses is to have an object that only lives in the server
but you can easily achieve this by setting an object as active
in the Started and Stopped events in NetworkServer.

In the future, we may reintroduce this if there is a good case for it
with an AOI rewrite

closes #389

BREAKING CHANGE: Remove serverOnly option in NetworkIdentity