Remove static NC.Current and NS.Current
As far as I know,  this is the last piece of static state we have.

Instead of having it as a static, which could cause trouble with
multiple tasks,  when we receive a message, we attach the current
NetworkClient and NetworkServer to NetworkReader.

A custom deserializer,  can ask the reader for the NC or NS and use
it to lookup NetworkIdentities.