feat: Transport client now collects stats
Now the KCP transport reuses code between client and server.
You can connect to multiple servers with the same socket.

Note that you cannot connect to the same server multiple time with
the same transport, as it would be interpreted as the same client.

If you want that,  then each client needs their own transport instance.