refactor: Rename PlayerSpawner to CharacterSpawner
As discussed in Discord,  Player is the meat bag between the keyboard
and the chair. The gameobject he controls is the Character.
There may be any number of characters from 0 to n per Player.

Therefore this class is renamed to stay consistent with the new
terminology.

BREAKING CHANGE: Renamed PlayerSpawner to CharacterSpawner