refactor: remove local visibility hacks
The NetworkVisibility disabled the renderers in host mode.
This is just a hack and it is unnecesary.

With this patch,  host mode just sees all objects.
If the user wants to limit how much the host player sees,
they can do it just as easily by changing the far clipping plane in the camera,
or implementing whatever logic they want with layers.

BREAKING CHANGE: NetworkVisibility no longer disables renderers in host mode