perf: serverrpc bypasses network on host mode (#708)
When calling a server rpc on server mode (whether host or not), there is no need to go through the network. Simply invoke the user code.

This changes behaviour a little. When you call a ServerRpc, it now executes immediately. Before, it executed on the next frame.

This also allows the ServerRpc to be invoked from the server

BREAKING CHANGE: ServerRpc execute synchronous in host mode