feat: precondition don't throw exceptions
This is an RFC about error handling and bugs.
The motivation from this came from a 2019 CppCon talk about exceptions in C++

The logic goes like this:

There are 3 types of errors:

1) Runtime errors that can happen at any time, such as null pointer
exceptions, stack overflow, out of memory errors, out of band access.
Generally we should not try to recover from these errors,  a human needs
to address them. We should never throw one of these errors ourselves.
We should have logic that prevent these error4s from happening.
For example, we should check for null when null is valid
and behave accordingly, so we would not see a NullReferenceException.

2) Plain bugs, typical case are when the user violates preconditions.
For example,  if the developer creates a NetworkClient with no transport
and starts it,  this is not a problem we should try to catch and recover
from. Instead we should report it to the developer as a bug.

3) Recoverable errors,  these are things that can happen during normal operation of the game, but happen
when a function is not able to complete successfully.  For example,  if we receive an invalid message from the network.

The goal of these PR is to make a distinction between 2 and 3.  And for 2, use Asserts instead of
throwing exceptions.

Note I just did a few so that we can have a discussion.
There are many more that should be turned into asserts too.