perf: improving how socketlayer handles endpoints (#856)
Peer EndPoint is now an interface that has a method to create a copy. This will allow Sockets to re-use a single endpoint inside receive and Peer will create a copy of it when a new connection is accepted

BREAKING CHANGE: Socket functions now use an interface instead of the EndPoint class, Socket Implementations should create a custom Endpoint class for their socket.