refactor: networkplayer sceneis ready notw defaults to true
When a player first connects it is now assumed that they are in a ready scene. NetworkSceneManager will set them to not ready if it has any scenes to load.

Spawning Visible objects will still wait for a character to be added before spawning.

BREAKING CHANGE: NetworkPlayer.SceneIsReady now default to true