perf: making transform values option to send with spawn message (#972)
Not all objects need their transform values synced when they are spawned.
And many object have it sent twice because the values will also be sent by NT.

This adds the option to disable syncing position for some Identities by disabling the option in the inspector. The values default to true so this will not break any existing projects.

This also makes the values in spawn message nullable so they will take up 3 bits extra when they are all synced. But this should not increase the otherall size because writing the payload will padd to the nearest byte. and there are already 2 bools in this struct