fix(NetworkWorld): fixing add identity when object is destroyed client side
if an object is destroyed locally client side it wont be removed from world, but if the server then respawns it then world will throw an error because netid already exists there. It should instead check if the existing object is null or not, If it is then it is ok to add a new one.