fix: adding more validate for invoking ServerRpc in host mode
Making serverRPC only be invoked if client and server are active locally.
Also checking if host client has auth (if required) so that the call behaviours same as in server client mode.

This change is to stop mistakes where ServerRpc is called in server only mode.
Instead a regular method should be used to share code for server and rpc method

BREAKING CHANGE: ServerRpc are now only invoked locally if in host mode