fix: adding null check for failing to resolve assembly
will give compiler warning instead of fail. It should be fine to fail some references.

It will just means Extension methods wont be found in them, but if reference isn't found it is unlikely to be a dll part of the same project (eg it could be system dll that couldn't be found)