fix: scene bug and GameObject scene move
When running disable commands, we have to make sure it only invokes when necessary. This means ignoring events such that the scene is registered in the forced visibility scenes.

As well, when GameObjects switch scenes, this will by default go undetected by the visibility manager. To accommodate for this, I added a CheckScenesUpdate boolean. While it's not performant, it will make sure visibility is always correct. For much better performance, you must manually invoke the UpdateGameObjectVisibility method with the GameObject who's scene was changed. This will update the renderer state internally and disable it if necessary.