fix: fixing syncvar not being set from another dll
Previously Syncvar would only be set from within the same DLL because the list of getter and setters are only checked within the dll

With this change we store the name of get/set in the attribute, so we can check that attribute from any dll we process.

Also an internal change, instead of properties. weaver now generate methods. It will generate GetNetwork__ only for wrapper syncvars and SetNetwork__ for all syncvars