feat(NetworkManagerGUI): overhaul the debug controls (#1149)
The current NetworkManagerGUI that is used by Mirage is rather outdated. Buttons are hard to click on if you have a touch screen, the white text blends in with the scene, etc. Overall, it was begging for an update.

This PR introduces an almost-completely refactored NetworkManagerGUI that is using the GUI.Window functionality to give it a nice background and title. In the future, one could turn it into a draggable GUI Window, but for now it's more elegant but keeping things simple.

Empty space will be used in later updates (current connection count, etc).