feat: adding syncList methods to set an index as dirty
when using classes with synclist there is no way to detect if a class is changed. and setting the index to be the same item again requires first setting it to null.

the new function allows us to use it instead of having to find the index, set null, then set value again.