feat: adding attribute that allows for combination of checks
Allows people to do a combination of condition with OR checks.

This is not possible with current attributes because they will be combine using AND, [Server,  HasAuthority] would require both server and hasAuthority to be true for the method to go away. Where as new [NetworkMethod(NetworkFlags.Server | NetworkFlags.HasAuthority)] would allow method to run if either server or HasAuthority to be true.