fix: Auth success message wording in NetworkClient (#1183)
This is a simple fix to stop the NetworkClient emitting a weird line that is "Authentication succcessful with" (where there's nothing past 'with' since message.AuthenticatorName is null/empty).
This commit now makes it emit either a success message with the name or it was successful w/o authenticator.

Chance of breaking existing projects? Unlikely.
Will it break existing tests? Checked and didn't see any tests checking the logging being emitted, so (safely?) assuming No.